# perf: reduce decoder malloc & reuse existing vars