--- a/decoder/decoder_test.go
+++ b/decoder/decoder_test.go
@@ -1404,3 +1404,47 @@ func TestReadValue(t *testing.T) {
 		})
 	}
 }
+
+func BenchmarkDecodeMessageData(b *testing.B) {
+	b.StopTimer()
+	mesg := factory.CreateMesg(typedef.MesgNumRecord)
+	mesgDef := proto.CreateMessageDefinition(&mesg)
+	dec := New(nil, WithIgnoreChecksum(), WithNoComponentExpansion())
+	dec.localMessageDefinitions[0] = &mesgDef
+	mesgb, _ := mesg.MarshalBinary()
+	b.StartTimer()
+
+	for i := 0; i < b.N; i++ {
+		dec.r = bytes.NewBuffer(mesgb)
+		err := dec.decodeMessageData(0)
+		if err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
+func BenchmarkDecode(b *testing.B) {
+	b.StopTimer()
+	// This is not a typical FIT in term of file size (2.3M) and the messages it contains (200.000 messages)
+	// But since it's big, it's should be good to benchmark.
+	f, err := os.Open("../testdata/big_activity.fit")
+	if err != nil {
+		panic(err)
+	}
+	defer f.Close()
+
+	all, err := io.ReadAll(f)
+	if err != nil {
+		panic(err)
+	}
+	b.StartTimer()
+
+	dec := New(nil)
+	for i := 0; i < b.N; i++ {
+		dec.r = bytes.NewBuffer(all)
+		for dec.Next() {
+			_, _ = dec.Decode()
+		}
+		dec.reset()
+	}
+}