- decoder: use single backing array instead of creating slice on each r read since the max size is known.
- decoder: implement reuse existing variables/instance rather than creating new one.
- proto: remove Size in FieldBase, we can get the size from basetype.
- profile: clean up fitgen's profile builder.go
- generate: profile and factory


benchmark:
```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/decoder
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
                    │    old.txt    │               new.txt                │
                    │    sec/op     │    sec/op     vs base                │
DecodeMessageData-4   11.317µ ± 13%   9.543µ ±  1%  -15.68% (p=0.000 n=10)
Decode-4               160.7m ± 44%   137.0m ± 11%  -14.76% (p=0.001 n=10)
geomean                1.349m         1.144m        -15.22%

                    │   old.txt    │               new.txt               │
                    │     B/op     │     B/op      vs base               │
DecodeMessageData-4   3.430Ki ± 0%   3.277Ki ± 0%  -4.47% (p=0.000 n=10)
Decode-4              97.52Mi ± 0%   94.85Mi ± 0%  -2.74% (p=0.000 n=10)
geomean               585.3Ki        564.2Ki       -3.61%

                    │   old.txt    │               new.txt               │
                    │  allocs/op   │  allocs/op   vs base                │
DecodeMessageData-4    136.00 ± 0%    58.00 ± 0%  -57.35% (p=0.000 n=10)
Decode-4              1900.1k ± 0%   900.1k ± 0%  -52.63% (p=0.000 n=10)
geomean                16.08k        7.225k       -55.05%

```